requirements.txt
scipy==1.1.0
faiss==1.5.3
numpy==1.19.5
scikit_image==0.17.2
huffman==0.1.2
opencv_python==4.4.0.46
Pillow==8.3.2
scikit_learn==0.24.2
skimage==0.0
